Dynamic sprite drawing
Changed the sprites to be drawn dynamically so changing the position should change where the sprite is drawn to. Fixed the bug where it would leave a trail of previously selected tiles. TODO fix new bug which occurs when the curser is moved over an area with a lot of sprites too quickly and often and causes them to bug out and all turn black.